Simplify SDK dev/dev-once auth flow: split registration from token exchange
The client_credentials grant requires a workspace Application to exist, but
token exchange was happening before createDevelopmentApplication. This caused
400 errors on first run or after uninstall.

- Split resolve-app-access-token.ts into focused files:
  ensure-app-registration.ts, ensure-app-access-token-is-valid-or-refresh.ts,
  exchange-credentials-for-tokens.ts, index.ts barrel
- Flatten dev-once pipeline: register -> install -> upload/sync -> token -> generateClient
- Remove token exchange from orchestrator initializePipeline, move to generate step
- Rename ensureValidAppAccessTokenOrRefresh to ensureAppAccessTokenIsValidOrRefresh
  with optional credentials fallback for client_credentials exchange

Made-with: Cursor
