+++ packages/twenty-sdk/src/cli/utilities/auth/index.ts
@@ -1,0 +1,3 @@
+export { ensureAppAccessTokenIsValidOrRefresh } from './ensure-app-access-token-is-valid-or-refresh';
+export { ensureAppRegistration } from './ensure-app-registration';
+export { exchangeCredentialsForTokens } from './exchange-credentials-for-tokens';
