+++ packages/twenty-sdk/src/cli/utilities/dev/orchestrator/steps/generate-api-client-orchestrator-step.ts
@@ -1,4 +1,4 @@
-import { ensureValidAppAccessTokenOrRefresh } from '@/cli/utilities/auth/resolve-app-access-token';
+import { ensureAppAccessTokenIsValidOrRefresh } from '@/cli/utilities/auth';
 import { type ClientService } from '@/cli/utilities/client/client-service';
 import { type ConfigService } from '@/cli/utilities/config/config-service';
 import { type OrchestratorState } from '@/cli/utilities/dev/orchestrator/dev-mode-orchestrator-state';
@@ -26,15 +26,19 @@ export class GenerateApiClientOrchestratorStep {
     this.notify = notify;
   }
 
-  async execute(input: { appPath: string }): Promise<void> {
+  async execute(input: {
+    appPath: string;
+    credentials?: { clientId: string; clientSecret: string };
+  }): Promise<void> {
     const step = this.state.steps.generateApiClient;
 
     step.status = 'in_progress';
     this.notify();
 
     try {
-      const appAccessToken = await ensureValidAppAccessTokenOrRefresh(
+      const appAccessToken = await ensureAppAccessTokenIsValidOrRefresh(
         this.configService,
+        input.credentials,
       );
 
       await this.clientService.generateCoreClient({
